requirements.txt
langchain==0.1.13
langchain-community==0.0.29
langchain-google-genai  # adjust based on your install
streamlit
python-dotenv==1.0.0
pypdf
tqdm==4.66.2
nest-asyncio==1.6.0
google-generativeai==0.5.0
pinecone
typing-extensions>=4.5.0
